Add files via upload
Here is the code - my idea was to log their password and connect with the website they use it with - then monitor their status if they are active in our system or not active...maybe we give them 30days of no activity then they turn not active - also, once a password expires and they make a new one we log that in the archives.  We can make a VIEW where user can look up past passwords already used.
I want to add file for Badpasswords - the thought there was to compare to inputted password, if it matches to one in the badpassword list we send an error to enter an new one.
Also for Dummy data - we go to filldb.info/dummy to fill.